==> code/1_RTC.ipynb <==
SOURCES DE DONNÉES & MÉTADONNÉES
---------------------------------------------------------
1. Données brutes (RTC) :
   https://www.donneesquebec.ca/recherche/dataset/rtc-gtfs-arrets-et-les-parcours

2. Documentation technique (GTFS Reference) :
   https://gtfs.org/documentation/schedule/reference/

---------------------------------------------------------
Consulté et téléchargé le : 11 janvier 2026

Parcours RTc : https://www.rtcquebec.ca/horaires-et-trajets/horaires-des-bus/tous-les-parcours

import pandas as pd
import os

# Lecture des données du RTC 
path = '../input/'

df_agency = pd.read_csv(path + 'agency.txt')
df_routes = pd.read_csv(path + 'routes.txt')
df_stops = pd.read_csv(path + 'stops.txt')
df_trips = pd.read_csv(path + 'trips.txt')
df_stop_times = pd.read_csv(path + 'stop_times.txt', low_memory=False);df_stop_times.columns = df_stop_times.columns.str.strip()
df_calendar = pd.read_csv(path + 'calendar.txt')
df_shapes = pd.read_csv(path + 'shapes.txt')
df_transfers = pd.read_csv(path + 'transfers.txt')

df_agency

df_routes.head()

df_stops.head()

df_trips.head()

df_stop_times.head()

df_calendar.head()

df_shapes.head()

df_transfers.head()

# Tables d'intéret

Ici je ne garde que les tables qui m'interessent (stops, trips et stop_times). Je veux relier les parcours (route_id) aux arrets (stop_id). 


Je fais une jointure entre df_routes et df_stop_time par trip_id  pour lier les voyages et les horaires

ensuite le resultats est joint a la table df_stop sur le stop_id pour recuperer les coordonnées et les nom d'arrets 

Source https://gtfs.org/documentation/schedule/reference/

# Je vais chercher lier les voyages et les horaires de passage
merged_step_1 = pd.merge(df_trips[['route_id',  'trip_id', 'direction_id', 'trip_headsign','wheelchair_accessible']],
                         df_stop_times[['trip_id', 'stop_id', 'stop_sequence', 'pickup_type',  'drop_off_type']], 
                         on='trip_id')

# Je rajoute les infos sur les arrets (lat, long, nom)
merged_step_2 = pd.merge(merged_step_1, df_stops, on='stop_id')

# Nettoyage finale et suppression des doublons
merged_arret_parcours = merged_step_2[[
    'route_id', 
    'direction_id', 
    'trip_headsign',
    'stop_id', 
    'stop_name', 
    'stop_lat', 
    'stop_lon', 
    'stop_sequence',
    'wheelchair_accessible', 
    'wheelchair_boarding',  
    'pickup_type',
    'drop_off_type'
]].drop_duplicates()

merged_arret_parcours['Parcours'] = merged_arret_parcours['route_id'].astype(str).str.split('-').str[-1] # route_id = 1-76 devient parcours = 76

merged_arret_parcours.head()

len(merged_arret_parcours.Parcours.unique())

# Parcours du RTC

import geopandas as gpd

df_parcours = gpd.read_file('../output/PArcoursJson/Parcours.json') # fichier crée sur arcgis pro a partir d'un shapefile 
df_parcours['Parcours'] = df_parcours['Parcours'].astype(str)
df_parcours = df_parcours.drop_duplicates(subset=['Parcours', 'Type'])
df_parcours.head()

len(df_parcours.Parcours.unique())

df_final = pd.merge(merged_arret_parcours, df_parcours[['Parcours', 'Type']], on='Parcours', how='left')
df_final.head(3)

df_final["Type"].unique()

# Je ne garde que les parcours de type 'Régulier' et 'Métrobus' et aussi uniquement les arrêts où la descente est permise (drop_off_type == 0)
df_final = df_final[
    (df_final['Type'].isin(['Régulier', 'Métrobus'])) & 
    (df_final['drop_off_type'] == 0)
]

df_final = df_final.drop(columns=['wheelchair_accessible', 'wheelchair_boarding', 'pickup_type', 'drop_off_type'])

gdf = gpd.GeoDataFrame(
    df_final, 
    geometry=gpd.points_from_xy(df_final.stop_lon, df_final.stop_lat),
    crs="EPSG:4326"  #  WGS84
)

gdf.crs

#gdf = gdf[gdf.Type == 'Régulier']

gdf.to_file("../output/rtc_data.geojson", driver='GeoJSON')
