==> code/4_MergeEverything.ipynb <==
import warnings
warnings.simplefilter("ignore")

import geopandas as gpd
import pandas as pd
import numpy as np

rtc_data = gpd.read_file("../output/rtc_data.geojson")
osm_places = gpd.read_file('../output/osm_places.geojson')
lieuxpublics = gpd.read_file('../input/DonneesVilleQC/vdq-lieupublic.geojson') 

# RTC

rtc_data.head()

# Nettoyage données OSM

osm_places = osm_places[~osm_places['amenity'].isin(['parking', 'bus_station'])] # je decide de ne pas les garder

# Fusion name et goepy_name (en priorisant name), si je n'ai pas de valeurs manquantes dans la colonne name j attribut cette valeur a geopy_name 
osm_places.loc[osm_places['name'].notna(), 'geopy_name'] = osm_places['name']

osm_places[osm_places.amenity == 'post_office'].name.unique()

osm_places.head()

mots_cles_adress = r'^(\d+|Rue|Avenue|Boulevard|Rang|Chemin|Route|Place|Autoroute)'

adresse = osm_places['geopy_name'].str.contains(mots_cles_adress, case=False, regex=True, na=False) # je verifie si le nom ressemble plus a une adresse qu a un nom de place 
nom_valide = osm_places['name'].notna() # si j ai un nom valide dans la colonne name et pas une valeur manquante

# Filtre final, on garde si (c'est un nom de place) OU SI (ce n'est pas une adresse)
place_to_keep = nom_valide | ~ adresse
osm_places = osm_places[place_to_keep]

pd.DataFrame(osm_places.groupby(['amenity']).size())

# Fusion OSM et données villde de QC ?

# Projection WGS84 vers le NAD83 MTM Zone 7
osm_places_m = osm_places.to_crs(epsg=32198)
lieuxpublics_m = lieuxpublics.to_crs(epsg=32198)

# Jointure spatiale dans un rayon de 10m : si deux points sont a moins de 10m, cela veut peut etre dire qu'il s'agit de la meme place  
df_merged = gpd.sjoin_nearest(
    osm_places_m, 
    lieuxpublics_m, 
    how="inner",        
    max_distance=10,    
    distance_col="distance_metres"
)
df_merged[['geopy_name', 'NOM_TOPOGRAPHIE', 'distance_metres']].head(30)

# Il y a bel et bien des doublons Centre culturel Georges Dor" et "Centre culturel Georges-Dor" , il s agit du meme batiment 
# Il y a aussi des mauvais match, "Centre culturel Georges Dor" et "Bureau d'arrondissement..." c'est juste un bureau situé dans le même bâtiment ou un batiment juste à côté

# Pistes / idées a tester
# Utiliser la comparaison textuelle pour supprimer les doublons dans les jeux de données en plus de la distance 
# Creer un score de similarité des textes et fusionner avec les distances pour decider s'il s'agit de la mme place ou non 
# Ce code est une ébauche, donc cette approche est à explorer plus tard

osm_places.to_file("../output/osm_places_v2.geojson", driver='GeoJSON')
